+++ frameworks/libxgame/src/lpeg/lpvm.c
@@ -77,10 +77,10 @@ static Capture *growcap (lua_State *L, Capture *capture, int *capsize,
 static Stack *doublestack (lua_State *L, Stack **stacklimit, int ptop) {
   Stack *stack = getstackbase(L, ptop);
   Stack *newstack;
-  int n = *stacklimit - stack;  /* current stack size */
+  int n = (int)(*stacklimit - stack);  /* current stack size */
   int max, newn;
   lua_getfield(L, LUA_REGISTRYINDEX, MAXSTACKIDX);
-  max = lua_tointeger(L, -1);  /* maximum allowed size */
+  max = (int)lua_tointeger(L, -1);  /* maximum allowed size */
   lua_pop(L, 1);
   if (n >= max)  /* already at maximum size? */
     luaL_error(L, "backtrack stack overflow (current limit is %d)", max);
@@ -115,7 +115,7 @@ static int resdyncaptures (lua_State *L, int fr, int curr, int limit) {
       luaL_error(L, "invalid position returned by match-time capture");
   }
   lua_remove(L, fr);  /* remove first result (offset) */
-  return res;
+  return (int)res;
 }
 
 
@@ -313,7 +313,7 @@ const char *match (lua_State *L, const char *o, const char *s, const char *e,
         captop -= n;  /* remove nested captures */
         ndyncap -= rem;  /* update number of dynamic captures */
         fr -= rem;  /* 'rem' items were popped from Lua stack */
-        res = resdyncaptures(L, fr, s - o, e - o);  /* get result */
+        res = resdyncaptures(L, fr, (int)(s - o), (int)(e - o));  /* get result */
         if (res == -1)  /* fail? */
           goto fail;
         s = o + res;  /* else update current position */
