+++ frameworks/libxgame/src/sproto/lsproto.c
@@ -259,7 +259,7 @@ encode(const struct sproto_arg *args) {
 			return SPROTO_CB_ERROR;
 		memcpy(args->value, str, sz);
 		lua_pop(L,1);
-		return sz;
+		return (int)sz;
 	}
 	case SPROTO_TSTRUCT: {
 		struct encode_ud sub;
@@ -313,7 +313,7 @@ static int
 lencode(lua_State *L) {
 	struct encode_ud self;
 	void * buffer = lua_touserdata(L, lua_upvalueindex(1));
-	int sz = lua_tointeger(L, lua_upvalueindex(2));
+	int sz = (int)lua_tointeger(L, lua_upvalueindex(2));
 	int tbl_index = 2;
 	struct sproto_type * st = lua_touserdata(L, 1);
 	if (st == NULL) {
@@ -397,7 +397,7 @@ decode(const struct sproto_arg *args) {
 		break;
 	}
 	case SPROTO_TBOOLEAN: {
-		int v = *(uint64_t*)args->value;
+		int v = (int)(*(uint64_t*)args->value);
 		lua_pushboolean(L,v);
 		break;
 	}
@@ -541,11 +541,11 @@ lpack(lua_State *L) {
 	size_t maxsz = (sz + 2047) / 2048 * 2 + sz + 2;
 	void * output = lua_touserdata(L, lua_upvalueindex(1));
 	int bytes;
-	int osz = lua_tointeger(L, lua_upvalueindex(2));
+	int osz = (int)lua_tointeger(L, lua_upvalueindex(2));
 	if (osz < maxsz) {
-		output = expand_buffer(L, osz, maxsz);
+		output = expand_buffer(L, osz, (int)maxsz);
 	}
-	bytes = sproto_pack(buffer, sz, output, maxsz);
+	bytes = sproto_pack(buffer, (int)sz, output, (int)maxsz);
 	if (bytes > maxsz) {
 		return luaL_error(L, "packing error, return size = %d", bytes);
 	}
@@ -559,13 +559,13 @@ lunpack(lua_State *L) {
 	size_t sz=0;
 	const void * buffer = getbuffer(L, 1, &sz);
 	void * output = lua_touserdata(L, lua_upvalueindex(1));
-	int osz = lua_tointeger(L, lua_upvalueindex(2));
-	int r = sproto_unpack(buffer, sz, output, osz);
+	int osz = (int)lua_tointeger(L, lua_upvalueindex(2));
+	int r = sproto_unpack(buffer,(int) sz, output, osz);
 	if (r < 0)
 		return luaL_error(L, "Invalid unpack stream");
 	if (r > osz) {
 		output = expand_buffer(L, osz, r);
-		r = sproto_unpack(buffer, sz, output, r);
+		r = sproto_unpack(buffer, (int)sz, output, r);
 		if (r < 0)
 			return luaL_error(L, "Invalid unpack stream");
 	}
@@ -594,7 +594,7 @@ lprotocol(lua_State *L) {
 	t = lua_type(L,2);
 	if (t == LUA_TNUMBER) {
 		const char * name;
-		tag = lua_tointeger(L, 2);
+		tag = (int)lua_tointeger(L, 2);
 		name = sproto_protoname(sp, tag);
 		if (name == NULL)
 			return 0;
@@ -636,7 +636,7 @@ static struct sproto * G_sproto[MAX_GLOBALSPROTO];
 static int
 lsaveproto(lua_State *L) {
 	struct sproto * sp = lua_touserdata(L, 1);
-	int index = luaL_optinteger(L, 2, 0);
+	int index = (int)luaL_optinteger(L, 2, 0);
 	if (index < 0 || index >= MAX_GLOBALSPROTO) {
 		return luaL_error(L, "Invalid global slot index %d", index);
 	}
@@ -647,7 +647,7 @@ lsaveproto(lua_State *L) {
 
 static int
 lloadproto(lua_State *L) {
-	int index = luaL_optinteger(L, 1, 0);
+	int index = (int)luaL_optinteger(L, 1, 0);
 	struct sproto * sp;
 	if (index < 0 || index >= MAX_GLOBALSPROTO) {
 		return luaL_error(L, "Invalid global slot index %d", index);
